PF-971-8-5-2-change-apis
Dearest Reviewer,

Rails 5.1 and 5.2 change the changes api in a breaking way. This appears
to be the only call that is impacted.

Updated per the DEPRECATION WARNING!
DEPRECATION WARNING: The behavior of changed_attributes inside of after
callbacks will be changing in the next version of Rails. The new return
value will reflect the behavior of calling the method after save returned
(e.g. the opposite of what it returns now). To maintain the current behavior,
use saved_changes.transform_values(&:first) instead.

Becker